Use the proper size for the bytes pointer
* In `mono_unicode_from_external`, the `bytes` pointer should be a `glong`.
* The `mono_unicode_from_external` function passes `bytes` to `g_utf8_to_utf16` and casts it as a `glong*`.
* On 64-bit builds, `sizeof(gsize) == 8` and `sizeof(glong) == 4`, so the high four bytes of the `bytes` value were not being set.
* The only option for client code is to initialize the value passed to the `bytes` argument to a value of zero, which is not clearly necessary.
* Instead, let's change the type of the argument to `glong`, so that we pass something which won't end up with uninitialized data.

+++ mono/utils/strenc.c
@@ -41,7 +41,7 @@ static const char trailingBytesForUTF8[256] = {
  * of bytes in the returned string, not including the terminator.
  */
 gunichar2 *
-mono_unicode_from_external (const gchar *in, gsize *bytes)
+mono_unicode_from_external (const gchar *in, glong *bytes)
 {
 	gchar *res=NULL;
 	gchar **encodings;
